+++ CHANGELOG.md
@@ -7,10 +7,16 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+- Support for running builds on [CircleCI](https://circleci.com/).
+- A set of unit tests through [Buttercup](https://github.com/jorgenschaefer/emacs-buttercup).
+
+### Fixed
+- Do not set markers after rendering the test report buffer.
+
 ## [1.0.0] - 2018-11-13
 
 ### Added
-
 - Initial version with a set of features for running Midje tests, viewing
   reports and formatting tabular facts within Emacs.